# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/projects/": {
            "get": {
                "description": "Get a list of projects associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get projects associated with the authenticated user",
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "schema": {
                            "$ref": "#/definitions/controls.Result"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch projects",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new project and assign members to it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Create a new project",
                "parameters": [
                    {
                        "description": "Project details",
                        "name": "projectInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controls.UserEnterData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created project details",
                        "schema": {
                            "$ref": "#/definitions/controls.Result2"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create project",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects/{projectId}": {
            "put": {
                "description": "Edit the name of a project with the specified ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Edit a project's name",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the project to edit",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New project name",
                        "name": "projectInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controls.InputData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated the project",
                        "schema": {
                            "$ref": "#/definitions/controls.EditedProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID or request data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update project",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a project with the specified ID, including its members and bugs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the project to delete",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion confirmation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID or project ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/users": {
            "get": {
                "description": "Get a list of all users except the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users except the authenticated user",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "$ref": "#/definitions/controls.UsersDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch users",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controls.EditedProjectResponse": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                },
                "Updated Project": {
                    "$ref": "#/definitions/controls.Project"
                }
            }
        },
        "controls.InputData": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "controls.Project": {
            "type": "object",
            "properties": {
                "Created At": {
                    "type": "string"
                },
                "Project ID": {
                    "type": "integer"
                },
                "Project Name": {
                    "type": "string"
                },
                "Updated At": {
                    "type": "string"
                }
            }
        },
        "controls.Result": {
            "type": "object",
            "properties": {
                "bugID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByID": {
                    "type": "integer"
                },
                "createdByUserName": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "joinedAt": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "memberID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "memberUserID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "memberUserName": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "projectID": {
                    "type": "integer"
                },
                "projectName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "controls.Result2": {
            "type": "object",
            "properties": {
                "bugID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "createdByUserName": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "joinedAt": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "memberID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "memberUserID": {
                    "description": "Rename this field to match the column name",
                    "type": "integer"
                },
                "memberUserName": {
                    "description": "Rename this field to match the column name",
                    "type": "string"
                },
                "projectID": {
                    "type": "integer"
                },
                "projectName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "controls.UserEnterData": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controls.UsersDTO": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
    }
}